LPS-90208 Move to new packages
